[Refactor/#81] Refactor: 아이콘 타입 및 속성을 Icon 컴포넌트로 이동
